Bug #1387033 Infinite loop & Logical loop
[root cause  ] Infinite loop & Logical loop
[changes     ] Modify the variable range & Remove the judgment condition
[side effects]no
[self test   ]:no
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:SharkL3
[test case]:Coverity code scanning
[reviewers   ] hongcheng.Li alec.Zhu
[change_type ] feature_bugfix
[tag_product ] specific

+++ oem2v4/src/cmr_scale.c
@@ -108,7 +108,7 @@ int32_t cmr_scaling_down(struct img_frm *src, struct img_frm *dst) {
     cmr_u32 cur_byte = 0;
     cmr_u32 ratio_w;
     cmr_u32 ratio_h;
-    uint16_t i, j;
+    uint32_t i, j;
     if (NULL == dst || NULL == src) {
         return -1;
     }
